+++ lib/actor/success.rb
@@ -2,5 +2,5 @@
 
 class Actor
   # Raised when using `succeed!` to halt the progression of an organizer.
-  class Success < StandardError; end
+  class Success < Actor::Error; end
 end